Removed the static_cast's for the KeyCodes

+++ Hazel/src/Hazel/Events/KeyEvent.h
@@ -29,7 +29,7 @@ namespace Hazel {
 		std::string ToString() const override
 		{
 			std::stringstream ss;
-			ss << "KeyPressedEvent: " << static_cast<uint32_t>(m_KeyCode) << " (" << m_RepeatCount << " repeats)";
+			ss << "KeyPressedEvent: " << m_KeyCode << " (" << m_RepeatCount << " repeats)";
 			return ss.str();
 		}
 
@@ -47,7 +47,7 @@ namespace Hazel {
 		std::string ToString() const override
 		{
 			std::stringstream ss;
-			ss << "KeyReleasedEvent: " << static_cast<int32_t>(m_KeyCode);
+			ss << "KeyReleasedEvent: " << m_KeyCode;
 			return ss.str();
 		}
 
@@ -63,7 +63,7 @@ namespace Hazel {
 		std::string ToString() const override
 		{
 			std::stringstream ss;
-			ss << "KeyTypedEvent: " << static_cast<int32_t>(m_KeyCode);
+			ss << "KeyTypedEvent: " << m_KeyCode;
 			return ss.str();
 		}
